Complete module restructuring and optimization
REORGANIZATION:
- Restructured modules into logical hierarchy:
  - core/: config, app_controller, data_service, rtc_manager
  - hardware/: camera, led, pins, voltage_sensor
  - communication/: esp_now, network_manager
  - power/: sleep management
  - Removed unused files and consolidated responsibilities

IMPROVEMENTS:
- Removed unused deep_sleep_new.rs file
- Eliminated redundant functionality between data_collector and camera modules
- Created unified DataService for data collection and transmission
- Added proper CameraPins structure for pin management
- Fixed all import paths and API compatibility issues
- Added missing methods (MacAddress::new, StatusLed::blink_success)

COMPILATION:
- All code compiles successfully with only minor unused code warnings
- Fixed ownership issues and API mismatches
- Maintained functional compatibility while improving structure
