+++ devices/m5stack_unit_cam/src/communication/mod.rs
@@ -1,0 +1,5 @@
+/// 通信関連モジュール
+pub mod esp_now;
+pub mod network_manager;
+
+pub use network_manager::NetworkManager;
