+++ devices/m5stack_unit_cam/src/core/mod.rs
@@ -1,0 +1,10 @@
+/// コアシステムモジュール
+pub mod app_controller;
+pub mod config;
+pub mod data_service;
+pub mod rtc_manager;
+
+pub use app_controller::AppController;
+pub use config::{AppConfig, ConfigError};
+pub use data_service::{DataService, MeasuredData};
+pub use rtc_manager::RtcManager;
