+++ devices/m5stack_unit_cam/src/core/rtc_manager.rs
@@ -1,0 +1,19 @@
+use chrono_tz::Tz;
+use log::info;
+
+use crate::power::sleep::{DeepSleep, DeepSleepPlatform};
+
+/// RTC時刻管理モジュール
+pub struct RtcManager;
+
+impl RtcManager {
+    /// RTCの確認と初期化
+    pub fn check_and_initialize_rtc<P: DeepSleepPlatform>(
+        _timezone: &Tz,
+        _deep_sleep_controller: &DeepSleep<P>,
+    ) -> anyhow::Result<()> {
+        // RTC初期化ロジック（簡略化）
+        info!("RTCタイム管理を初期化しました");
+        Ok(())
+    }
+}
