+++ devices/m5stack_unit_cam/src/hardware/mod.rs
@@ -1,0 +1,8 @@
+/// ハードウェア制御モジュール
+pub mod camera;
+pub mod led;
+pub mod pins;
+pub mod voltage_sensor;
+
+pub use pins::CameraPins;
+pub use voltage_sensor::VoltageSensor;
