+++ devices/m5stack_unit_cam/src/hardware/voltage_sensor.rs
