+++ devices/m5stack_unit_cam/src/power/mod.rs
@@ -1,0 +1,4 @@
+/// 電源管理モジュール
+pub mod sleep;
+
+pub use sleep::{DeepSleep, DeepSleepError, EspIdfDeepSleep};
